criar dockerfile e publicar
